Product page
git-svn-id: https://sp-b2cv1.googlecode.com/svn/shopJoyBuy@1416 c8aa77e7-4d14-485d-9416-cb5e655923dd

+++ src/com/spshop/cache/SCacheFacade.java
@@ -186,7 +186,7 @@ public class SCacheFacade{
 	}
 	
 	
-	public static List<String> getTabProductNames(int tabProductId){
+	public static List<String> getTabProductNames(long tabProductId){
 		List<String> names = (List<String>) getTabProductNameCache().get(""+tabProductId);
 		if(null==names){
 			names = ServiceFactory.getService(TabProductService.class).getProductNames(tabProductId);
